distinguish komodo_check_deposit into separate check fn
